Add router to front-end
