+++ frontend/src/index.js
@@ -1,38 +1,12 @@
-import ReactDOM from 'react-dom/client'
+
 import React from 'react'
+import ReactDOM from 'react-dom';
 
-import Grid from '@mui/material/Grid';
-import Box from '@mui/material/Box';
+import App from './App'
 
-import ResponsiveAppBar from './ResponsiveAppBar'
-import { Event } from './Event';
-import { ForumOverview } from './ForumOverview';
-import { ArtItem } from './ArtItem';
-
-class HomePage extends React.Component {
-    render() {
-        return (
-            <Box >
-                <Box>
-                    <ResponsiveAppBar />
-                </Box>
-                <Grid container sx={{
-                    gap: 2
-                }}>
-                    <Grid item xs={7}>
-                        <Event />
-                        <ArtItem />
-                        <Event />
-                    </Grid>
-                    <Grid item xs>
-                        <ForumOverview />
-                    </Grid>
-                </Grid>
-            </Box>
-        )
-    }
-}
-
-const container = document.querySelector('#root')
-const root = ReactDOM.createRoot(container)
-root.render(<HomePage />)
+ReactDOM.render(
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>,
+    document.getElementById('root')
+);
