+++ frontend/src/pages/HomePage/ArtItem.js
